src/dspygen/experiments/done/openai_ror_cli.py

import typer
app = typer.Typer()

@app.command(name="create")
def create():
    """Create a new OpenAI project with the given name"""
    # Command logic goes here
    print("This is the create command.")

@app.command(name="train")
def train():
    """Train the OpenAI project with the given project ID"""
    # Command logic goes here
    print("This is the train command.")

@app.command(name="generate")
def generate():
    """Generate text based on the given prompt"""
    # Command logic goes here
    print("This is the generate command.")

@app.command(name="list")
def list():
    """List all OpenAI projects"""
    # Command logic goes here
    print("This is the list command.")

@app.command(name="delete")
def delete():
    """Delete the OpenAI project with the given project ID"""
    # Command logic goes here
    print("This is the delete command.")

@app.command(name="help")
def help():
    """Display help for a specific command"""
    # Command logic goes here
    print("This is the help command.")


if __name__ == "__main__":
    app()

